refactor: move to helpers
