+++ packages/api-reference/src/features/api-client-modal/ApiClientModal.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { useActiveEntities, useWorkspace } from '@scalar/api-client/store'
 import { mutateSecuritySchemeDiff } from '@scalar/api-client/views/Request/libs'
-import { getServersFromDocument } from '@scalar/oas-utils/transforms'
+import { getServersFromDocument } from '@scalar/oas-utils/helpers'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { ApiClientPlugin } from '@scalar/types/api-client'
 import type {
