+++ packages/oas-utils/src/helpers/index.ts
@@ -8,6 +8,7 @@ export { shouldUseProxy, redirectToProxy } from './redirect-to-proxy'
 export { schemaModel } from './schema-model'
 export { shouldIgnoreEntity } from './should-ignore-entity'
 export { isOperationDeprecated, getOperationStability, getOperationStabilityColor } from './operation-stability'
+export { getServersFromDocument } from './servers'
 
 /**
  * @deprecated These helpers are being phased out. Please import directly from @scalar/helpers instead.
