+++ packages/oas-utils/src/transforms/index.ts
@@ -1,7 +1,6 @@
 export {
   getSlugUid,
   importSpecToWorkspace,
-  getServersFromDocument,
   parseSchema,
   type ImportSpecToWorkspaceArgs,
 } from './import-spec'
